Fix Handling for B9PS Switching
Current configurations for B9PS do not include complete ModuleWaterfallFX config nodes, but only groups of either templates, controllers, and/or effects.  Separate out the handling for each element and replace nodes of that type in the main config as appropriate.
TODO: No implementation yet for controller switching, only effect and template.