Fix ramp rates and event durations -previously, ramp speed and event time was updated *every time* the controller was polled, which means the effective speed depended on how many effects were using the controller -with the performance optimizations in this branch, each controller only updates once per frame -the end result is that all ramp speeds and event durations come out wrong
We don't want to invalidate all existing content, so this change adds a versioning system to the controllers where they can adjust their values if they were saved with an older version.  When the effect module is re-saved it will be marked with the newer version and have the corrected values written out.